# docs/swagger.yaml
basePath: /
host: localhost:5000
info:
  contact:
    name: API Support
  description: swagger Projeto Raul
  title: Fiber Example API
  version: "1.0"
paths:
  /asset:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Asset
  /assets:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Assets
  /building:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Building
  /buildings:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Building
  /character:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Character
  /characters:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Characters
  /climate:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Climate
  /climates:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Climates
...
  /deleteAsset/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Asset by its ID
      operationId: delete-Asset-by-int
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Asset
      tags:
      - Asset
  /deleteBuilding/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Building by its ID
      operationId: delete-building-by-int
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Building
      tags:
      - Item
  /deleteCharacter/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Character by its ID
      operationId: delete-Character-by-int
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Character
      tags:
      - Item
  /deleteClimate/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Climate by its ID
      operationId: delete-Climate-by-int
      parameters:
      - description: Climate ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Climate
      tags:
      - Item
...
  /updateAsset/{id}:
    put:
      consumes:
      - application/json
      description: update a Asset by its ID
      operationId: get-Asset-by-int
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Asset
      tags:
      - Asset
  /updateBuilding/{id}:
    put:
      consumes:
      - application/json
      description: update a Building by its ID
      operationId: get-building-by-int
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Building
      tags:
      - Building
  /updateCharacter/{id}:
    put:
      consumes:
      - application/json
      description: update a Character by its ID
      operationId: get-Character-by-int
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Character
      tags:
      - Item
  /updateClimate/{id}:
    put:
      consumes:
      - application/json
      description: update a Climate by its ID
      operationId: get-climate-by-int
      parameters:
      - description: Climate ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Climate
      tags:
      - Item
...
swagger: "2.0"